테스트 클래스에서 공통적으로 쓰일 필드와 메서드를 슈퍼 클래스로 추출하라
